Refactoring as per Joerg's comments and aligning the package name to the service name
