+++ bundle/src/main/java/com/adobe/acs/commons/filefetch/FileFetcher.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,18 +17,22 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.fetchfile;
+package com.adobe.acs.commons.filefetch;
 
 import java.io.IOException;
 
 import com.day.cq.replication.ReplicationException;
 
 /**
- * OSGi Service for
+ * OSGi Service for fetching files from a remote source and caching them in AEM Assets.
  */
-public interface FileFetch {
-  
+public interface FileFetcher {
+
+  FileFetchConfiguration getConfig();
+
   Exception getLastException();
+
+  String getLastModified();
 
   boolean isLastJobSucceeded();
